technicaltest2/API/requirements.txt
Flask>=3.0
PyJWT>=2.8.0
flask-cors>=4.0
Werkzeug>=3.0
